Create Project 1 Source Files
Q1 Write a Java program that calculates and displays the Fibonacci
series, defined by the recursive formula F(n) = F(n-1) + F(n-2), and a method not using recursion. 

Q2. Write a Hello World Greeter program, 

Q3. Write a program that:
a. reads from the terminal a sequence of numbers (integers)
b. saves them to a file with the name given from the command line
c. calculates, then displays on the terminal, and also saves to that file
the maximum, minimum, and average.
Store the numbers in a LinkedList<Integer>.
Define a class DataAnalyzer that
 * has a constructor that stores the list of numbers:
public DataAnalyzer(LinkedList<Integer> numList) {...}
 * has a method each for computing min(), max() and average():
public int min() {...}, etc.
Define a class DataAnalyzerTester that reads the numbers from System.in,
builds the number list,
creates the DataAnalyzer object, and displays the min, max, and average
using the DataAnalyzer instance.
The DataAnalyzerTester class implements the main() method.
Your code needs to handle invalid input and I/O exceptions.

Q5 Write a program that computes the prime factorization of a positive
integer number. The number in question is
computed as the product of prime numbers raised to various exponents. The
prime factorization problem is
finding the primes and their exponents.